- changed BGP identifier to type long to avoid int signing problems

+++ lib/common-config/src/main/java/org/bgp4/config/nodes/PeerConfiguration.java
@@ -54,13 +54,13 @@ public interface PeerConfiguration {
 	 * 
 	 * @return
 	 */
-	public int getRemoteBgpIdentifier();
+	public long getRemoteBgpIdentifier();
 	
 	/**
 	 * 
 	 * @return
 	 */
-	public int getLocalBgpIdentifier();
+	public long getLocalBgpIdentifier();
 	
 	/**
 	 *
